# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tracepad API",
    "summary": "Lightweight, self-hosted, OTLP-native store for LLM and agent traces.",
    "description": "The read API is the product's contract: the UI, the CLI and the MCP server are thin clients over exactly these endpoints, and a feature that is not here does not exist. Responses are flat snake_case JSON with RFC 3339 timestamps and opaque cursors. Every response spends a byte budget on payloads (default 51200, `?budget=` to override): a payload that does not fit is cut at a UTF-8 boundary and replaced by a truncation marker naming where the whole of it lives.",
    "license": {"name": "Apache-2.0", "identifier": "Apache-2.0"},
    "version": "0.1.0"
  },
  "servers": [{"url": "http://localhost:4318", "description": "A locally running tracepad"}],
  "security": [{"bearerKey": []}, {"basicKeys": []}],
...
  "paths": {
...
    "/api/v1/sessions": {
      "get": {
        "tags": ["traces"],
        "summary": "List sessions by most recent activity, filtered and cursor-paginated",
        "description": "Aggregated from the traces that named a session; there is no session table. `from`/`to` bound the *traces*, half-open like the trace listing, so a session appears when any of its traces falls in the window and its totals then describe those traces. To read one session's traces, follow with `GET /api/v1/sessions/{id}`.",
        "operationId": "listSessions",
        "parameters": [
          {"$ref": "#/components/parameters/From"},
          {"$ref": "#/components/parameters/To"},
          {"$ref": "#/components/parameters/Environment"},
          {"name": "user_id", "in": "query", "schema": {"type": "string"},
            "description": "Exact match on the user id of a session's traces"},
          {"$ref": "#/components/parameters/Limit"},
          {"$ref": "#/components/parameters/Cursor"},
          {"$ref": "#/components/parameters/Direction"},
          {"$ref": "#/components/parameters/Count"}
        ],
        "responses": {
          "200": {"description": "A page of sessions", "content": {"application/json": {"schema": {
            "type": "object",
            "properties": {
              "sessions": {"type": "array", "items": {"$ref": "#/components/schemas/SessionRow"}},
              "next_cursor": {"type": ["string", "null"], "description": "Pass back as `?cursor=` for the next page; null on the oldest one"},
              "prev_cursor": {"type": ["string", "null"], "description": "Pass back with `?direction=prev` for the page before; null on the newest one"},
              "total": {"type": "integer", "description": "Present only with `?count=`: how many sessions the filters match, capped at 1000"},
              "total_capped": {"type": "boolean", "description": "Present only with `?count=`: the count stopped at the cap"}
            },
            "required": ["sessions", "next_cursor", "prev_cursor"]
          }}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
...
    "/api/v1/stats": {
      "get": {
        "tags": ["stats"],
        "summary": "Counts, errors, cost and latency percentiles per bucket",
        "description": "`unit` says what a bucket counts: grouping by hour, day, environment, release or total counts traces, and grouping by model counts observations, because a trace has no model. Grouped by release, a trace that named none falls in the bucket with the empty key. Grouped by `total`, the whole window is one bucket under the empty key — the same shape as a day bucket, with percentiles merged over every hour in the window — and it takes every filter the other groupings take. An empty range answers with no buckets rather than with fabricated zero rows. Latency percentiles are histogram-based: accurate to a few percent, and stable across the expiry of the raw rows.",
        "operationId": "stats",
        "parameters": [
          {"$ref": "#/components/parameters/From"},
          {"$ref": "#/components/parameters/To"},
          {"$ref": "#/components/parameters/Environment"},
          {"name": "user_id", "in": "query", "schema": {"type": "string"},
            "description": "Restricts every bucket to one end user. The shape, the groupings and `unit` are unchanged; a `hour` or `day` timeline, and the `total` bucket, additionally carry `sessions` per bucket. Statistics for one user trail the raw data by the same rollup lag as the rest"},
          {"name": "group_by", "in": "query",
            "schema": {"type": "string", "enum": ["hour", "day", "model", "environment", "release", "total"], "default": "day"}}
        ],
        "responses": {
          "200": {"description": "The buckets", "content": {"application/json": {"schema": {
            "type": "object",
            "properties": {
              "group_by": {"type": "string", "enum": ["hour", "day", "model", "environment", "release", "total"]},
              "unit": {"type": "string", "enum": ["trace", "observation"]},
              "buckets": {"type": "array", "items": {
                "type": "object",
                "properties": {
                  "key": {"type": "string"},
                  "count": {"type": "integer"},
                  "error_count": {"type": "integer"},
                  "total_cost": {"type": "number", "description": "Absent when nothing in the bucket carried a cost"},
                  "tokens": {"type": "object",
                    "description": "Token sums over the observations in the bucket, on every grouping and both units. Each key is present only when something in the bucket carried that count, and the object is absent when none of the three is — absent, never zero, like `total_cost` — and always absent with `user_id`, whose rollup holds no token sums. Cache-read tokens are the input tokens a provider reported as served from its cache; reasoning and cache-creation counts are not summed",
                    "properties": {
                      "input": {"type": "integer"},
                      "output": {"type": "integer"},
                      "cache_read": {"type": "integer"}
                    }},
                  "sessions": {"type": "integer",
                    "description": "Present only with `user_id` and an `hour`, `day` or `total` grouping: how many of that user's sessions began in this bucket. A session is counted where it starts, so a sum over any range is exact"},
                  "latency_ms": {"type": "object", "properties": {
                    "p50": {"type": ["integer", "null"]},
                    "p95": {"type": ["integer", "null"]}
                  }}
                },
                "required": ["key", "count", "error_count", "latency_ms"]
              }}
            },
            "required": ["group_by", "unit", "buckets"]
          }}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerKey": {"type": "http", "scheme": "bearer",
        "description": "A project's secret key, `tp-sk-…`"},
      "basicKeys": {"type": "http", "scheme": "basic",
        "description": "`base64(public key:secret key)`, the shape a Langfuse SDK sends"},
      "adminToken": {"type": "http", "scheme": "bearer",
        "description": "`TRACEPAD_ADMIN_TOKEN`, sent in the same header as a project key. Required for everything cross-project, for creating and deleting a project, and for the account routes; beside an owner's or editor's session, it is what lists, mints and revokes a project's keys, which no project key does — which is what makes it the recovery when every owner's password is lost. A server with none configured answers 403 and says so."},
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "tracepad_session",
        "description": "A browser session: `HttpOnly`, `SameSite=Lax`, `Secure` over TLS, set by the three ways in and lasting `TRACEPAD_SESSION_DAYS` (default 30), sliding. On a route whose policy is `member` or `editor` a session must also name the project it is asking about with `X-Tracepad-Project`, and an unsafe method must carry an `Origin` (or `Referer`) whose host is the request's `Host`, the first value of `X-Forwarded-Host`, or the host of `TRACEPAD_URL`."}
    },
    "parameters": {
      "From": {"name": "from", "in": "query", "schema": {"type": "string", "format": "date-time"},
        "description": "RFC 3339, inclusive. The range is half-open, so walking a timeline never reports a row twice."},
      "To": {"name": "to", "in": "query", "schema": {"type": "string", "format": "date-time"},
        "description": "RFC 3339, exclusive"},
      "Environment": {"name": "environment", "in": "query", "schema": {"type": "string"},
        "description": "The environment a trace ran in. A comma-separated list matches **any** of them: `?environment=production,staging`. Items are trimmed and duplicates collapse; an empty item (`a,,b`, `a,`) is a 400, as an empty value is, and so is repeating the parameter — one list, one parameter. At most 100 items; beyond that, a 400. An environment whose name contains a comma is not expressible here. `GET /api/v1/facets` lists the values in a range with their counts"},
      "Limit": {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 500, "default": 50},
        "description": "Out of range is a 400, not a silent clamp"},
      "Cursor": {"name": "cursor", "in": "query", "schema": {"type": "string"},
        "description": "The opaque `next_cursor` or `prev_cursor` of a previous page"},
      "Direction": {"name": "direction", "in": "query", "schema": {"type": "string", "enum": ["next", "prev"], "default": "next"},
        "description": "Which way to page from the cursor. With no cursor, `next` is the newest page and `prev` the oldest — both ends are a direction rather than an offset. Rows come back newest first either way"},
      "Count": {"name": "count", "in": "query", "schema": {"type": "string", "enum": ["1", "true"]},
        "description": "Adds `total` and `total_capped`: how many rows the filters match, counted up to 1000. Off by default, because the count changes with the filters and not with the page"},
...
    },
    "responses": {
      "BadRequest": {"description": "The request cannot mean what it says: an unknown query parameter, a malformed value, a limit out of range",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "Unauthorized": {"description": "The credentials do not resolve to a project",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "ServiceUnavailable": {"description": "The credentials could not be checked right now, more passwords were being checked at once than the server allows, or a write met a database condition that passes (a full disk, a lock that did not clear), or a read found no free read slot before its deadline (`the server is busy; retry shortly`): retry after `Retry-After`. A read the deadline stopped while it ran (`the read took longer than 20s and was stopped; narrow the time range or the filters`, the number being `TRACEPAD_READ_TIMEOUT`) carries no `Retry-After`: the same request would be stopped again, so narrow it",
        "headers": {"Retry-After": {"schema": {"type": "string"}, "description": "Seconds to wait"}},
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The one error shape the whole surface uses",
        "properties": {"error": {"type": "string"}},
        "required": ["error"]
      },
...
      "SessionRow": {
        "type": "object",
        "description": "The roll-up over the traces of one session. Every number counts traces, not observations: `error_count` is how many of the session's traces failed.",
        "properties": {
          "id": {"type": "string", "description": "The session id, as the application set it"},
          "trace_count": {"type": "integer"},
          "error_count": {"type": "integer"},
          "total_cost": {"type": "number", "description": "Summed over the traces that carried a cost; absent when none did"},
          "first_seen": {"type": "string", "format": "date-time"},
          "last_seen": {"type": "string", "format": "date-time"}
        },
        "required": ["id", "trace_count", "error_count"]
      },
...
    }
  }
}